Initial training loop
Wrote initial training loop

+++ src/training/dataloader.py
@@ -54,7 +54,7 @@ class ASRDataset(Dataset):
 
         return (
             padded_mel_filter_bank,
-            mel_mask.long(),
+            mel_mask,
             tokenized.long(),
             token_mask.long(),
         )
